modify pb for wxvm

+++ module/vm/wxvm/runtime.go
@@ -65,7 +65,7 @@ func (r *RuntimeInstance) Invoke(contractId *commonPb.ContractId, method string,
 		contractResult.Message = err.Error()
 		return
 	} else {
-		contractResult.GasUsed = int64(inst.ExecCtx.GasUsed())
+		contractResult.GasUsed = inst.ExecCtx.GasUsed()
 		contractResult.ContractEvent = context.ContractEvent
 	}
